chore(types): extract suppressions types into @monokle/types
